# optimize `inflate::core::transfer`