@oyvindln 2023-01-09
Really nice.

Yeah lack of safe SIMD is an issue I guess. Often the compiler is smart enough to use it but it varies. rustc and underlying llvm have some options to help debug why it didn't vectorize stuff but it's clunky.